+++ BetterPrismUI/_Thunderstore/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-Nov-16
+
+### Added
+
+- Prism health is now displayed for other players in the Party UI
+- Prism health is now displayed for healthbars shown above players in the world
+
 ## [1.0.0] - 2025-Nov-01
 
 ### Changed